+++ src/main/java/org/visuals/legacy/animatium/mixins/v1/rendering/MixinMinecraft_ColorBoost.java
@@ -1,0 +1,79 @@
+/**
+ * Animatium
+ * The all-you-could-want legacy animations mod for modern minecraft versions.
+ * Brings back animations from the 1.7/1.8 era and more.
+ * <p>
+ * Copyright (C) 2024-2025 lowercasebtw
+ * Copyright (C) 2024-2025 mixces
+ * Copyright (C) 2024-2025 Contributors to the project retain their copyright
+ * <p>
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * <p>
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * <p>
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * <p>
+ * "MINECRAFT" LINKING EXCEPTION TO THE GPL
+ */
+
+package org.visuals.legacy.animatium.mixins.v1.rendering;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.mojang.blaze3d.pipeline.RenderPipeline;
+import com.mojang.blaze3d.pipeline.RenderTarget;
+import com.mojang.blaze3d.systems.RenderPass;
+import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.DefaultVertexFormat;
+import com.mojang.blaze3d.vertex.VertexFormat;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.renderer.RenderPipelines;
+import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
+import org.spongepowered.asm.mixin.injection.At;
+import org.visuals.legacy.animatium.Animatium;
+import org.visuals.legacy.animatium.config.AnimatiumConfig;
+
+import java.util.OptionalInt;
+
+@Mixin(Minecraft.class)
+public abstract class MixinMinecraft_ColorBoost {
+	@Shadow
+	@Nullable
+	public Screen screen;
+
+	@Unique
+	private static final RenderPipeline animatium$boostPipeline = RenderPipelines.register(
+			RenderPipeline.builder()
+					.withLocation(Animatium.location("pipeline/colorboost"))
+					.withVertexShader("core/screenquad")
+					.withFragmentShader(Animatium.location("core/colorboost"))
+					.withSampler("Sampler0")
+					.withVertexFormat(DefaultVertexFormat.EMPTY, VertexFormat.Mode.TRIANGLES)
+					.build()
+	);
+
+	@WrapOperation(method = "runTick(Z)V", at = @At(value = "INVOKE", target = "Lcom/mojang/blaze3d/pipeline/RenderTarget;blitToScreen()V"))
+	private void animatium$colorBoost(final RenderTarget instance, final Operation<Void> original) {
+		if (AnimatiumConfig.instance().extras.colorBoost && this.screen == null) {
+			try (final RenderPass renderPass = RenderSystem.getDevice().createCommandEncoder().createRenderPass(() -> "Color boost render target blit", instance.getColorTextureView(), OptionalInt.empty())) {
+				renderPass.setPipeline(animatium$boostPipeline);
+				RenderSystem.bindDefaultUniforms(renderPass);
+				renderPass.bindSampler("Sampler0", instance.getColorTextureView());
+				renderPass.draw(0, 3);
+			}
+		}
+
+		original.call(instance);
+	}
+}