+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Old Water Color Effects
 - Fix thrown egg kicking you
 - Option to turn on/off 1.7 "longer un-sneak" animation
+- New "extras" setting to color-boost the game
 - Singleplayer toggles for server features (in extras config tab)
 
 ## Changed: